C prototipo/BackEnd General/blockchain-processor/requirements.txt
# Blockchain Processor - GASLYT
# Dependencias Python para procesamiento de datos blockchain

# Blockchain
web3==6.13.0
eth-account==0.9.0
cryptography==41.0.7

# Base de datos
pymysql==1.1.0
influxdb-client==1.38.0

# Configuración y logging
python-dotenv==1.0.0
structlog==23.2.0

# Utilidades
requests==2.31.0
schedule==1.2.0
tenacity==8.2.3

# Métricas y estadísticas
pandas==2.1.3
numpy==1.25.2

# Desarrollo
pytest==7.4.3
black==23.11.0
flake8==6.1.0
